anim update - 01
Created basic structure a project demonstrating react and css transitions. Added 3 different transition demonstrations.  One shows basic transition of a single element with a background.  The  second demonstrates using CSS transitions to animate multiple states on button click.  The last is a component that adds and subtracts a list of elements on button click and adds transitions when they are added/removed.