nba 03
Added an articles component.  The component demonstrates react useEffect and useState to get a list of articles from the database and display them in a grid.   Added a poll component that allows user to modify a counter for different teams.  The component displays the teams and an image icon for each team.  Added a subscribe component that allows a user to subscribe using an email.  The component saves the input email to the database and also checks if it already exists and displays appropriate error messages or returns the thank you message.