redux 3
Created getMoviesList function that grabs a list of posts from a web api.  Created functional component that demonstrates redux mapStateToProps and mapDispatchToProps functions to display the movies list on click.  Also demonstrates using redux connect to link the helper functions with the functional component.  Added hooks component that uses redux hooks methods useDispatch and useSelector to get and update the redux store state.  The component also uses hooks useEffect to run the dispatch function and get the movies list.  Added redux-promise middleware wrapper to the react store to handle promises and wait for resolved status to return the data