forms 01
Added base structure for forms project.  Added formData template to formOne component.  The config contains data on element fields and events used for validation and value updating.  Added formFields component that renders the template based on the formData prop and handles error messages and validation.  The component uses the template to display an input and update the state on change.  Added validation function that checks if the element contains requires validation and if so returns true and the validation message.  This is used in the error handling in the formFields component.