forms 02
Added lastname age and message objects to template.  Added onSuccess reset function that resets the form value and validation statuses.  Added submitForm method that loops through each property in the formData and checks it's isValid state.  If the form is valid it loops and updates each properties value to the new element input value.  Added FormField components to the main form which renders and handles actions for each form element.  Added validate case that checks if the min age from the input is >= the required age minimum and if so displays an error message