Bookshelf 17
Fixed bug where once a user was deleted accessing their posts would throw an error.  This was because the ownerId for the deleted user was being checked in the singleBook route and didn't exist.  Updated the component to just display the current user.  Added css updates to changes colors and hover effects based on new color scheme.  Fixed bug where my posts were displaying all posts as the wrong ownerId was being checked in the all posts route.  Added screenshots and comments.